requirements.txt
Pillow
pystray
